chore: update imports following restructure
