+++ api.planx.uk/modules/analytics/analyticsLog/service.ts
@@ -1,5 +1,5 @@
 import { gql } from "graphql-request";
-import { $public } from "../../client/index.js";
+import { $public } from "../../../client/index.js";
 
 interface UpdateAnalyticsLogUserExit {
   analyticsLog: {
